+++ CHANGELOG.md
@@ -4,9 +4,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### Added
+
+- Added methods `hash` and `rawHash` for byte arrays
+- Added method `pbkdf` to ease generating key material from a password
+
 ### Changed
 
-- Marked all methods accepting a `String` password as deprecated. These methods will be removed in version 3.
+- Marked all methods accepting a `String` password as deprecated. These methods will be removed in version 3
 
 ## [2.5] - 2018-10-07