+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 - Marked all methods accepting a `String` password as deprecated. These methods will be removed in version 3
 - Update Argon2 libraries to 20190702
+- Update JNA to 5.4.0
 
 ## [2.5] - 2018-10-07